+++ src/compositions/song_2.rs
@@ -1,5 +1,8 @@
 use event::{Event, Mutate, Phrase, Render};
-use instrument::oscillator::{Oscillator, StereoWaveform};
+use instrument::{
+    oscillator::Oscillator,
+    stereo_waveform::StereoWaveform
+};
 use ratios::{Pan, R};
 use settings::get_default_app_settings;
