+++ src/instrument/mod.rs
@@ -1,5 +1,6 @@
 pub mod oscillator;
 pub mod voice;
+pub mod stereo_waveform;
 mod loudness;
 #[cfg(test)]
 mod test;
