+++ src/write.rs
@@ -1,5 +1,5 @@
 extern crate hound;
-use instrument::oscillator::StereoWaveform;
+use instrument::stereo_waveform::StereoWaveform;
 
 pub fn write_output_buffer(out_buffer: &mut [f32], stereo_waveform: StereoWaveform) {
     let mut l_idx = 0;
